cloud_function/monitoring_pubsub_gcs_to_bq/requirements.txt
# Function dependencies, for example:
# package>=version

# Import Librairy
google-cloud-bigquery
google-cloud-storage
